+++ src/controllers/inputsListener.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include <functional>
 
-void handleInputs(std::function<void(double)> loop);
+void handleInputs(std::function<void(int)> loop);
 
 void onKeyPressed(int code);
