+++ src/domains/gameData.hpp
@@ -6,5 +6,5 @@ struct GameData
 {
     GameData();
     std::vector<std::pair<int, int>> activePiece;
-    double delaySinceGravity;
+    int delaySinceGravity;
 };
